+++ test/mix_unused/analyzers/unused_test.exs
@@ -8,14 +8,14 @@ defmodule MixUnused.Analyzers.UnusedTest do
   doctest @subject
 
   test "no functions" do
-    assert %{} == @subject.analyze(%{}, [])
+    assert %{} == @subject.analyze(%{}, %{})
   end
 
   test "called externally" do
     function = {Foo, :a, 1}
     calls = %{Bar => [{function, %{caller: {:b, 1}}}]}
 
-    assert %{} == @subject.analyze(calls, [{function, %Meta{}}])
+    assert %{} == @subject.analyze(calls, %{function => %Meta{}})
   end
 
   test "called internally and externally" do
@@ -26,20 +26,20 @@ defmodule MixUnused.Analyzers.UnusedTest do
       Bar => [{function, %{caller: {:b, 1}}}]
     }
 
-    assert %{} == @subject.analyze(calls, [{function, %Meta{}}])
+    assert %{} == @subject.analyze(calls, %{function => %Meta{}})
   end
 
   test "called only internally" do
     function = {Foo, :a, 1}
     calls = %{Foo => [{function, %{caller: {:b, 1}}}]}
 
-    assert %{} == @subject.analyze(calls, [{function, %Meta{}}])
+    assert %{} == @subject.analyze(calls, %{function => %Meta{}})
   end
 
   test "not called at all" do
     function = {Foo, :a, 1}
 
-    assert %{^function => _} = @subject.analyze(%{}, [{function, %Meta{}}])
+    assert %{^function => _} = @subject.analyze(%{}, %{function => %Meta{}})
   end
 
   test "functions with metadata `:export` set to true are ignored" do
@@ -48,9 +48,7 @@ defmodule MixUnused.Analyzers.UnusedTest do
     assert %{} ==
              @subject.analyze(
                %{},
-               [
-                 {function, %Meta{doc_meta: %{export: true}}}
-               ]
+               %{function => %Meta{doc_meta: %{export: true}}}
              )
   end
 
@@ -59,7 +57,7 @@ defmodule MixUnused.Analyzers.UnusedTest do
     function_b = {Foo, :b, 1}
 
     calls = %{
-      Foo => [{function_b, %{function: {:a, 1}}}]
+      Foo => [{function_b, %{caller: {:a, 1}}}]
     }
 
     assert %{
@@ -68,10 +66,18 @@ defmodule MixUnused.Analyzers.UnusedTest do
            } =
              @subject.analyze(
                calls,
-               [
-                 {function_a, %Meta{}},
-                 {function_b, %Meta{}}
-               ]
+               %{function_a => %Meta{}, function_b => %Meta{}}
              )
+  end
+
+  test "functions called with default arguments are not reported" do
+    function = {Foo, :a, 1}
+    functions = %{function => %Meta{doc_meta: %{defaults: 1}}}
+    calls = %{Foo => [{{Foo, :a, 0}, %{caller: {:b, 1}}}]}
+
+    assert %{} ==
+             @subject.analyze(calls, functions, %{
+               entrypoints: [{Foo, :b, 1}]
+             })
   end
 end